requirements.txt
fastapi
uvicorn
pydantic

#depreciated
google-generativeai==0.8.5

python-jose[cryptography]==3.5.0